✨ Update environment configurations and refactor models
- Add FORCE_DB_SEED variable to backend .env.example to control DB seeding.
- Move HealthResponse, MessageRequest, and MessageResponse models to backend/models.py for better organization.
- Update main.py to import moved models from backend/models.py.
- Add VITE_FORCE_DB_SEED variable to frontend .env.example for local development.
- Create new MainListPage.tsx component file in frontend/src/components/products.
