+++ .env.example
@@ -42,6 +42,10 @@ MONGODB_HOST=mongodb
 MONGODB_DATABASE=seoulminds_db
 MONGODB_URI=mongodb://<YOUR_MONGO_ADMIN_USER>:<YOUR_MONGO_ADMIN_PASS>@mongodb:27017/seoulminds_db?authSource=admin
 
+# When set to 1 (or true), forces the backend to run DB seeding even if the
+# environment appears to be production. 
+FORCE_DB_SEED=1
+
 # ============================================================================
 # Mongo Express Configuration
 # ============================================================================
