+++ frontend/.env.example
@@ -10,3 +10,7 @@ VITE_APP_DESCRIPTION=hackseoulminds website
 
 # API Timeout (milliseconds)
 VITE_API_TIMEOUT=10000
+
+# Optional: instruct backend to force DB seeding during local development
+# (0 or 1)
+VITE_FORCE_DB_SEED=1
